+++ src/components/[guild]/RolesByPlatform/components/RoleListItem/RoleListItem.tsx
@@ -1,46 +1,28 @@
 import {
-  AlertDialog,
-  AlertDialogBody,
-  AlertDialogContent,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogOverlay,
   Button,
   Collapse,
   GridItem,
   Heading,
   HStack,
   Icon,
-  IconButton,
   SimpleGrid,
   Spinner,
   Stack,
   Tag,
   Text,
-  useBreakpointValue,
-  useDisclosure,
   VStack,
   Wrap,
 } from "@chakra-ui/react"
 import { useWeb3React } from "@web3-react/core"
 import GuildLogo from "components/common/GuildLogo"
-import useDelete from "components/[guild]/EditForm/components/DeleteCard/hooks/useDelete"
 import useIsOwner from "components/[guild]/hooks/useIsOwner"
 import LogicDivider from "components/[guild]/LogicDivider"
 import RequirementCard from "components/[guild]/RequirementCard"
 import useLevelsAccess from "components/[guild]/RolesByPlatform/components/JoinButton/hooks/useLevelsAccess"
-import usePersonalSign from "hooks/usePersonalSign"
 import useRequirementLabels from "hooks/useRequirementLabels"
-import { useRouter } from "next/router"
-import {
-  CaretDown,
-  CaretUp,
-  Check,
-  PencilSimple,
-  TrashSimple,
-  X,
-} from "phosphor-react"
-import React, { useRef, useState } from "react"
+import dynamic from "next/dynamic"
+import { CaretDown, CaretUp, Check, X } from "phosphor-react"
+import React, { useState } from "react"
 import { Role } from "types"
 import AccessIndicator from "./components/AccessIndicator"
 
@@ -51,20 +33,12 @@ type Props = {
 const RoleListItem = ({ roleData }: Props): JSX.Element => {
   const { account } = useWeb3React()
   const isOwner = useIsOwner(account)
-  const router = useRouter()
-
-  const { isOpen, onOpen, onClose } = useDisclosure()
-  const cancelRef = useRef()
-  const transition = useBreakpointValue<any>({ base: "slideInBottom", sm: "scale" })
-  const { onSubmit, isLoading: isRoleDeleteLoading } = useDelete(
-    "role",
-    roleData?.id
-  )
-  const { isSigning } = usePersonalSign()
 
   const { hasAccess, error, isLoading } = useLevelsAccess([roleData.id])
   const requirements = useRequirementLabels(roleData.requirements)
   const [isRequirementsExpanded, setIsRequirementsExpanded] = useState(false)
+
+  const DynamicOwnerButtons = dynamic(() => import("./components/OwnerButtons"))
 
   return (
     <Stack
@@ -148,57 +122,7 @@ const RoleListItem = ({ roleData }: Props): JSX.Element => {
             <AccessIndicator label="No access" icon={X} />
           ))}
 
-        {isOwner && (
-          <>
-            <Stack direction={{ base: "row", md: "column" }}>
-              <IconButton
-                icon={<Icon as={PencilSimple} />}
-                size="sm"
-                rounded="full"
-                aria-label="Edit role"
-                onClick={() =>
-                  router.push(`/${router.query.guild}/edit/${roleData.id}`)
-                }
-              />
-              <IconButton
-                icon={<Icon as={TrashSimple} color="red.500" />}
-                size="sm"
-                rounded="full"
-                aria-label="Delete role"
-                onClick={onOpen}
-              />
-            </Stack>
-
-            <AlertDialog
-              motionPreset={transition}
-              leastDestructiveRef={cancelRef}
-              {...{ isOpen, onClose }}
-            >
-              <AlertDialogOverlay>
-                <AlertDialogContent>
-                  <AlertDialogHeader>Delete Role</AlertDialogHeader>
-                  <AlertDialogBody>
-                    <Text>Are you sure? You can't undo this action afterwards.</Text>
-                  </AlertDialogBody>
-                  <AlertDialogFooter>
-                    <Button ref={cancelRef} onClick={onClose}>
-                      Cancel
-                    </Button>
-                    <Button
-                      colorScheme="red"
-                      isLoading={isRoleDeleteLoading}
-                      loadingText={isSigning ? "Check your wallet" : "Deleting"}
-                      onClick={() => onSubmit()}
-                      ml={3}
-                    >
-                      Delete
-                    </Button>
-                  </AlertDialogFooter>
-                </AlertDialogContent>
-              </AlertDialogOverlay>
-            </AlertDialog>
-          </>
-        )}
+        {isOwner && <DynamicOwnerButtons roleData={roleData} />}
       </HStack>
     </Stack>
   )
